+++ apps/backoffice-v2/src/domains/workflows/fetchers.ts
@@ -6,6 +6,7 @@ import { handleZodError } from '@/common/utils/handle-zod-error/handle-zod-error
 import { WorkflowDefinitionByIdSchema } from '@/domains/workflow-definitions/fetchers';
 import { AmlSchema } from '@/lib/blocks/components/AmlBlock/utils/aml-adapter';
 import { ObjectWithIdSchema } from '@/lib/zod/utils/object-with-id/object-with-id';
+import { CollectionFlowStates } from '@ballerine/common';
 import qs from 'qs';
 import { deepCamelKeys } from 'string-ts';
 import { z } from 'zod';
@@ -115,18 +116,18 @@ export const BaseWorkflowByIdSchema = z.object({
       })
       .passthrough()
       .optional(),
-    flowConfig: z
+    collectionFlow: z
       .object({
-        stepsProgress: z
-          .record(
-            z.string(),
-            z.object({
-              // TODO Until backwards compatibility is handled
-              number: z.number().default(0),
-              isCompleted: z.boolean(),
-            }),
-          )
-          .or(z.undefined()),
+        config: z.object({
+          apiUrl: z.string().url(),
+          steps: z.array(z.object({ stateName: z.string(), orderNumber: z.number() })),
+        }),
+        state: z.object({
+          uiState: z.string(),
+          collectionFlowState: z.enum(Object.values(CollectionFlowStates) as [string, ...string[]]),
+          progress: z.record(z.string(), z.object({ isCompleted: z.boolean() })),
+        }),
+        additionalInformation: z.record(z.string(), z.unknown()).optional(),
       })
       .optional(),
     customData: z.record(z.string(), z.unknown()).optional(),
@@ -150,7 +151,7 @@ export const WorkflowByIdSchema = BaseWorkflowByIdSchema.extend({
         context: true,
       }).extend({
         context: BaseWorkflowByIdSchema.shape.context.omit({
-          flowConfig: true,
+          collectionFlow: true,
         }),
       }),
     )