+++ packages/eslint-config-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/eslint-config-react
 
+## 2.0.22
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/eslint-config@1.1.22
+
 ## 2.0.21
 
 ### Patch Changes