+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ballerine/workflow-node-sdk
 
+## 0.6.57
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.57
+
+## 0.6.56
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.56
+
 ## 0.6.55
 
 ### Patch Changes